+++ car-rental-office-core/src/main/java/hu/uni/miskolc/iit/service/RentManagementService.java
@@ -20,7 +20,7 @@ public interface RentManagementService {
      * @param id
      * @return Rent object
      * */
-    Rent getRentById(int id);
+    Rent getRentById(Long id);
 
     /**
      * A szűrési paramétereknek megfelelő kölcsönöket adja vissza
